requirements.txt
# Core Framework
streamlit==1.31.0
python-dotenv==1.0.0

# LLM Integration
anthropic==0.18.1
openai==1.12.0

# Document Processing
PyPDF2==3.0.1
pypdf==4.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# NLP Libraries
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
nltk==3.8.1
langdetect==1.0.9
indic-nlp-library==0.92

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Report Generation
reportlab==4.0.9
fpdf2==2.7.7

# Utilities
python-dateutil==2.8.2
regex==2023.12.25
tqdm==4.66.1
